Corrected folder structure and removed response model for appendcampaign
