--- LiveVox.NET/Models/Campaign/AppendCampaign/AppendCampaignResponse.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using LiveVox.NET.Models.Base;
-
-namespace LiveVox.NET.Models.Campaign.AppendCampaign
-{
-    public class AppendCampaignResponse : ILiveVoxResponse
-    {
-    }
-}
