+++ LiveVox.NET/Models/LiveVoxSerializerContext.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 using LiveVox.NET.Models.Campaign.AppendCampaign;
+using LiveVox.NET.Models.Campaign.Common;
 using LiveVox.NET.Models.Campaign.Enumerations;
 using LiveVox.NET.Models.Session;
 
@@ -10,7 +11,6 @@ namespace LiveVox.NET.Models
     [JsonSerializable(typeof(SessionLoginRequest))]
     [JsonSerializable(typeof(SessionLoginResponse))]
     [JsonSerializable(typeof(AppendCampaignRequest))]
-    [JsonSerializable(typeof(AppendCampaignResponse))]
     public partial class LiveVoxSerializerContext : JsonSerializerContext
     {
         // The source generator will fill in the necessary code here.
